refactor: 스터디룸 정보를 가져오는 메소드들의 클래스 분리
